Refactoring of FIR endpoints in backend
