+++ new-backend/.env
@@ -214,5 +214,8 @@ OPENAPI_ENABLE_RESPONSE_VALIDATION=false
 #ANALYTICS_OPTION_TRACKERURL=https://xxxxxxxxxx.matomo.cloud
 #ANALYTICS_OPTION_SITEID=4
 
+# *** FIR SETTINGS ***
+# The url is an endpoint in myCarta FR Direkt
+# The output dir is where the generated xlsx files will be placed.
 # FIR_REALESTATEOWNERREPORT_URL=https://fastighetsrapport-utv.domain.se/Report/fastighetsforteckning/xlsx
 # FIR_TEMP_OUTPUT_DIR=/temp/
