+++ new-backend/server/apis/v2/router.js
@@ -5,7 +5,7 @@ import mapconfigRouter from "./controllers/mapconfig/router.js";
 import settingsRouter from "./controllers/settings/router.js";
 import informativeRouter from "./controllers/informative/router.js";
 import adRouter from "./controllers/ad/router.js";
-import fir from "./controllers/fir/router.js";
+import firRouter from "./controllers/fir/router.js";
 
 export default express
   .Router()
@@ -14,4 +14,4 @@ export default express
   .use("/mapconfig", mapconfigRouter)
   .use("/settings", settingsRouter)
   .use("/ad", adRouter)
-  .use("/fir", fir);
+  .use("/fir", firRouter);
